requirements.txt
streamlit==0.90.0
PyPDF2==1.26.0
google-generativeai==0.1.5
pymongo==3.12.0
